QPID-5653: Replace DurableConfigurationStore/MessageStore recoverers with visitors.
* MS/DCS impls now have stateless visitXXX methods to retrieve message/configuration data (replaces the recoverXXXX methods)
* VH implementations now uses Handlers to perform the recovery operation.
* DCS's handler (ConfiguredObjectRecordRecoveverAndUpgrader) currently implemented in terms of the old DefaultUpgradeProvider/DurableConfigurationRecoverer.
  This will be refactored by a future commit.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@1584600 13f79535-47bb-0310-9956-ffa450edef68

+++ java/broker-plugins/derby-store/src/test/java/org/apache/qpid/server/store/derby/DerbyMessageStoreTest.java
@@ -70,10 +70,13 @@ public class DerbyMessageStoreTest extends MessageStoreTestCase
 
     private void deleteStoreIfExists()
     {
-        File location = new File(_storeLocation);
-        if (location.exists())
+        if (_storeLocation != null)
         {
-            FileUtils.delete(location, true);
+            File location = new File(_storeLocation);
+            if (location.exists())
+            {
+                FileUtils.delete(location, true);
+            }
         }
     }
